+++ src/basics/doubly-linked-list/index.ts
@@ -1,0 +1,1 @@
+export * from './doubly-linked-list';
