+++ src/basics/linked-list/index.ts
@@ -1,0 +1,2 @@
+export * from './linked-list';
+export * from './array-like-linked-list';
