+++ utils/tests/lists.ts
@@ -5,6 +5,10 @@ interface ListBase<T> {
   next: ListBase<T> | null;
 }
 
+/**
+ * Gets the length of a list without built-in list methods
+ * @param listBase List.
+ */
 const getListLength = <T, TList extends ListBase<T>>(listBase: TList): number => {
   let length = 0;
   let current: ListBase<T> | null = listBase;
@@ -17,6 +21,11 @@ const getListLength = <T, TList extends ListBase<T>>(listBase: TList): number =>
   return length;
 }
 
+/**
+ * Gets a random element of the list.
+ * If the list size allows, the item is guaranteed not to be an outermost item.
+ * @param listBase List.
+ */
 export const getRandomListItem = <T, TList extends ListBase<T>>(listBase: TList): TList => {
   const listLength = getListLength(listBase);
   let from = 1;
